Remove diag_backup control logic to backup.inc
